rename multi_panel_plot to multi_inst_plot
